backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
yfinance
passlib[bcrypt]
python-jose[cryptography]
python-multipart
psycopg2-binary
